separate setProtocolOptions into utils.py
